+++ src/screens/MissionScreen/index.tsx
@@ -1,13 +1,21 @@
-import {View, StyleSheet, ImageBackground, Platform} from 'react-native';
-import React from 'react';
+import {
+  View,
+  StyleSheet,
+  ImageBackground,
+  Platform,
+  TouchableOpacity,
+  Image,
+} from 'react-native';
+import React, {useEffect} from 'react';
 
 import {BottomTabNavigationProp} from '@react-navigation/bottom-tabs';
 import {TabNavigatorParamList} from '../../navigations/bottomTabs/MainTapNavigator';
 import {colors, font, image} from '../../assets';
 import Text from '../../components/Text';
-import TabCustom from '../../components/TabCustom/TabCustom';
 import Body from './Body';
 import {SafeAreaView} from 'react-native-safe-area-context';
+import {Campaign} from '../../datasource/CampaignDatasource';
+import Spinner from 'react-native-loading-spinner-overlay';
 
 type MissionScreenProps = {
   navigation: BottomTabNavigationProp<TabNavigatorParamList, 'mission'>;
@@ -36,6 +44,25 @@ export default function MissionScreen({navigation}: MissionScreenProps) {
   const handleTabChange = (value: string) => {
     setSelectedTab(value);
   };
+  const [campaignImage, setCampaignImage] = React.useState<string>('');
+  const [loading, setLoading] = React.useState(false);
+
+  const fetchImage = async () => {
+    await Campaign.getImage('DRONER', 'QUATA', 'ACTIVE')
+      .then(res => {
+        setLoading(true);
+        setCampaignImage(res.data[0].pathImageFloating);
+      })
+      .catch(err => console.log(err))
+      .finally(() => {
+        setLoading(false);
+      });
+  };
+  useEffect(() => {
+    if (navigation.isFocused()) {
+      fetchImage();
+    }
+  }, [navigation]);
 
   return (
     <SafeAreaView
@@ -69,7 +96,27 @@ export default function MissionScreen({navigation}: MissionScreenProps) {
           /> */}
         </View>
       </ImageBackground>
-      <Body navigation={navigation} />
+      <Body navigation={navigation} fetchImage={fetchImage} />
+      <View
+        style={{
+          position: 'absolute',
+          bottom: 16,
+          right: 8,
+          zIndex: 1,
+        }}>
+        <TouchableOpacity onPress={() => navigation.navigate('CampaignScreen')}>
+          <Image
+            source={{uri: campaignImage}}
+            style={{width: 150, height: 60}}
+            resizeMode="contain"
+          />
+        </TouchableOpacity>
+      </View>
+      <Spinner
+        visible={loading}
+        textContent="Loading..."
+        textStyle={{color: '#FFF'}}
+      />
     </SafeAreaView>
   );
 }